+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Syslog configuration is now no longer configured using boolean in the log
+  structure but rather by using functions with default handled the same way as
+  for standard error fallback output.
+
+
 ## [0.3.0] - 2021-05-22
 ### Added 
 - support for binding logs to allow combination of multiple log instances to